+++ src/video_core/regs_framebuffer.h
@@ -193,7 +193,7 @@ struct FramebufferRegs {
         case ColorFormat::RGBA4:
             return 2;
         default:
-            LOG_CRITICAL(HW_GPU, "Unknown color format %u", format);
+            LOG_CRITICAL(HW_GPU, "Unknown color format %u", static_cast<u32>(format));
             UNIMPLEMENTED();
         }
     }
@@ -258,7 +258,7 @@ struct FramebufferRegs {
             return 4;
         }
 
-        ASSERT_MSG(false, "Unknown depth format %u", format);
+        ASSERT_MSG(false, "Unknown depth format %u", static_cast<u32>(format));
     }
 
     // Returns the number of bits per depth component of the specified depth format
@@ -271,7 +271,7 @@ struct FramebufferRegs {
             return 24;
         }
 
-        ASSERT_MSG(false, "Unknown depth format %u", format);
+        ASSERT_MSG(false, "Unknown depth format %u", static_cast<u32>(format));
     }
 
     INSERT_PADDING_WORDS(0x20);
