+++ src/video_core/renderer_opengl/gl_rasterizer.cpp
@@ -1338,7 +1338,8 @@ void RasterizerOpenGL::SyncCullMode() {
         break;
 
     default:
-        LOG_CRITICAL(Render_OpenGL, "Unknown cull mode %d", regs.rasterizer.cull_mode.Value());
+        LOG_CRITICAL(Render_OpenGL, "Unknown cull mode %u",
+                     static_cast<u32>(regs.rasterizer.cull_mode.Value()));
         UNIMPLEMENTED();
         break;
     }
